feat: Consolidate fade state management into audio store and remove fade store
