+++ ui/src/components/AudioTrackPlayer.vue
@@ -7,11 +7,9 @@
 import Hls from 'hls.js';
 import { onBeforeUnmount, onMounted, ref, watch } from 'vue';
 import { useAudioStore, type AudioTrack } from '../stores/audio';
-import { useFadeStore } from '../stores/fadeStore';
 
 const props = defineProps<{ fileID: string }>()
 const audioStore = useAudioStore()
-const fadeStore = useFadeStore()
 const videoElement = ref<HTMLVideoElement | null>(null)
 
 interface FadeState {
@@ -83,7 +81,7 @@ function startFade(startVolume: number, targetVolume: number) {
     targetVolume,
     startTime: Date.now()
   }
-  fadeStore.setFading(props.fileID, true)
+  audioStore.setFading(props.fileID, true)
 }
 
 function applyFadeStep(track: AudioTrack, videoEl: HTMLVideoElement) {
@@ -96,7 +94,7 @@ function applyFadeStep(track: AudioTrack, videoEl: HTMLVideoElement) {
 
   if (fadePercent === 1) {
     fadeState.value = null
-    fadeStore.setFading(track.fileID, false)
+    audioStore.setFading(track.fileID, false)
     if (currentVolume <= 0.01) {
       videoEl.pause()
     }
